requirements.txt
PyYAML~=6.0
tweepy~=4.14.0